Restructuring the tests
