+++ test/src/basic_api_test_helper.hpp
@@ -99,273 +99,3 @@ inline void run_test(uint32_t symbols, uint32_t symbol_size)
 }
 
 
-template<class Encoder, class Decoder>
-inline void
-invoke_basic_api(uint32_t symbols, uint32_t symbol_size)
-{
-
-    // Common setting
-    typename Encoder::factory encoder_factory(symbols, symbol_size);
-    auto encoder = encoder_factory.build();
-
-    typename Decoder::factory decoder_factory(symbols, symbol_size);
-    auto decoder = decoder_factory.build();
-
-    EXPECT_TRUE(symbols == encoder_factory.max_symbols());
-    EXPECT_TRUE(symbol_size == encoder_factory.max_symbol_size());
-    EXPECT_TRUE(symbols == encoder->symbols());
-    EXPECT_TRUE(symbol_size == encoder->symbol_size());
-
-    EXPECT_TRUE(symbols == decoder_factory.max_symbols());
-    EXPECT_TRUE(symbol_size == decoder_factory.max_symbol_size());
-    EXPECT_TRUE(symbols == decoder->symbols());
-    EXPECT_TRUE(symbol_size == decoder->symbol_size());
-
-    EXPECT_TRUE(encoder->symbol_length() > 0);
-    EXPECT_TRUE(decoder->symbol_length() > 0);
-
-    EXPECT_TRUE(encoder->block_size() == symbols * symbol_size);
-    EXPECT_TRUE(decoder->block_size() == symbols * symbol_size);
-
-    EXPECT_TRUE(encoder_factory.max_payload_size() >=
-                encoder->payload_size());
-
-    EXPECT_TRUE(decoder_factory.max_payload_size() >=
-                decoder->payload_size());
-
-    EXPECT_EQ(encoder_factory.max_payload_size(),
-              decoder_factory.max_payload_size());
-
-    // Encode/decode operations
-    EXPECT_EQ(encoder->payload_size(), decoder->payload_size());
-
-    std::vector<uint8_t> payload(encoder->payload_size());
-
-    std::vector<uint8_t> data_in = random_vector(encoder->block_size());
-    std::vector<uint8_t> data_in_copy(data_in);
-
-    sak::mutable_storage storage_in = sak::storage(data_in);
-    sak::mutable_storage storage_in_copy = sak::storage(data_in_copy);
-
-    EXPECT_TRUE(sak::equal(storage_in, storage_in_copy));
-
-    // Only used for prime fields, lets reconsider how we implement
-    // this less intrusive
-    uint32_t prefix = 0;
-
-    if(fifi::is_prime2325<typename Encoder::field_type>::value)
-    {
-        // This field only works for multiple of uint32_t
-        assert((encoder->block_size() % 4) == 0);
-
-        uint32_t block_length = encoder->block_size() / 4;
-
-        fifi::prime2325_binary_search search(block_length);
-        prefix = search.find_prefix(storage_in_copy);
-
-        // Apply the negated prefix
-        fifi::apply_prefix(storage_in_copy, ~prefix);
-    }
-
-    encoder->set_symbols(storage_in_copy);
-
-    // Set the encoder non-systematic
-    if(kodo::is_systematic_encoder(encoder))
-        kodo::set_systematic_off(encoder);
-
-    while( !decoder->is_complete() )
-    {
-        uint32_t payload_used = encoder->encode( &payload[0] );
-        EXPECT_TRUE(payload_used <= encoder->payload_size());
-
-        decoder->decode( &payload[0] );
-    }
-
-    std::vector<uint8_t> data_out(decoder->block_size(), '\0');
-    decoder->copy_symbols(sak::storage(data_out));
-
-    if(fifi::is_prime2325<typename Encoder::field_type>::value)
-    {
-        // Now we have to apply the negated prefix to the decoded data
-        fifi::apply_prefix(sak::storage(data_out), ~prefix);
-    }
-
-    EXPECT_TRUE(std::equal(data_out.begin(),
-                           data_out.end(),
-                           data_in.begin()));
-}
-
-template<class Encoder, class Decoder>
-inline void
-invoke_out_of_order_raw(uint32_t symbols, uint32_t symbol_size)
-{
-    // Common setting
-    typename Encoder::factory encoder_factory(symbols, symbol_size);
-    auto encoder = encoder_factory.build();
-
-    typename Decoder::factory decoder_factory(symbols, symbol_size);
-    auto decoder = decoder_factory.build();
-
-    // Encode/decode operations
-    EXPECT_TRUE(encoder->payload_size() == decoder->payload_size());
-
-    std::vector<uint8_t> payload(encoder->payload_size());
-    std::vector<uint8_t> data_in = random_vector(encoder->block_size());
-
-    encoder->set_symbols(sak::storage(data_in));
-
-    if(kodo::is_systematic_encoder(encoder))
-        kodo::set_systematic_off(encoder);
-
-    while( !decoder->is_complete() )
-    {
-
-        if((rand() % 100) > 50)
-        {
-            encoder->encode( &payload[0] );
-            decoder->decode( &payload[0] );
-        }
-        else
-        {
-            uint32_t symbol_id = rand() % encoder->symbols();
-
-            if(decoder->symbol_pivot(symbol_id))
-            {
-                continue;
-            }
-
-            ASSERT_TRUE(encoder->symbol_size() <= payload.size());
-
-            encoder->copy_symbol(
-                symbol_id, sak::storage(payload));
-
-            const uint8_t *symbol_src = encoder->symbol(symbol_id);
-            const uint8_t *symbol_dest = &payload[0];
-
-            EXPECT_TRUE(std::equal(symbol_src,
-                                   symbol_src + encoder->symbol_size(),
-                                   symbol_dest));
-
-
-            decoder->decode_symbol(&payload[0], symbol_id);
-
-        }
-    }
-
-    EXPECT_EQ(encoder->block_size(), decoder->block_size());
-
-    std::vector<uint8_t> data_out(decoder->block_size(), '\0');
-    decoder->copy_symbols(sak::storage(data_out));
-
-    EXPECT_TRUE(std::equal(data_out.begin(),
-                           data_out.end(),
-                           data_in.begin()));
-
-}
-
-/// Checks that the encoders and decoders are in a clean state after using
-/// the initialize function.
-template<class Encoder, class Decoder>
-inline void
-invoke_initialize(uint32_t symbols, uint32_t symbol_size)
-{
-
-    // Common setting
-    typename Encoder::factory encoder_factory(symbols, symbol_size);
-    auto encoder = encoder_factory.build();
-
-    typename Decoder::factory decoder_factory(symbols, symbol_size);
-    auto decoder = decoder_factory.build();
-
-
-    for(uint32_t i = 0; i < 10; ++i)
-    {
-        encoder->initialize(encoder_factory);
-        decoder->initialize(decoder_factory);
-
-        std::vector<uint8_t> payload(encoder->payload_size());
-
-        // Ensure that the we may re-use the encoder also with partial
-        // data.
-        uint32_t block_size = rand_nonzero(encoder->block_size());
-
-        std::vector<uint8_t> data_in = random_vector(block_size);
-
-        encoder->set_symbols(sak::storage(data_in));
-
-        // Set the encoder non-systematic
-        if(kodo::is_systematic_encoder(encoder))
-            kodo::set_systematic_off(encoder);
-
-        while( !decoder->is_complete() )
-        {
-            uint32_t payload_used = encoder->encode( &payload[0] );
-            EXPECT_TRUE(payload_used <= encoder->payload_size());
-
-            decoder->decode( &payload[0] );
-        }
-
-        std::vector<uint8_t> data_out(block_size, '\0');
-        decoder->copy_symbols(sak::storage(data_out));
-
-        bool data_equal = sak::equal(sak::storage(data_out),
-                                     sak::storage(data_in));
-
-        ASSERT_TRUE(data_equal);
-
-    }
-
-}
-
-
-
-template<class Encoder, class Decoder>
-inline void
-invoke_systematic(uint32_t symbols, uint32_t symbol_size)
-{
-
-    // Common setting
-    typename Encoder::factory encoder_factory(symbols, symbol_size);
-    auto encoder = encoder_factory.build();
-
-    typename Decoder::factory decoder_factory(symbols, symbol_size);
-    auto decoder = decoder_factory.build();
-
-    // Encode/decode operations
-    EXPECT_TRUE(encoder->payload_size() == decoder->payload_size());
-
-    std::vector<uint8_t> payload(encoder->payload_size());
-    std::vector<uint8_t> data_in = random_vector(encoder->block_size());
-
-    encoder->set_symbols(sak::storage(data_in));
-
-    // Ensure encoder systematic
-    EXPECT_TRUE(kodo::is_systematic_encoder(encoder));
-    kodo::set_systematic_on(encoder);
-
-    uint32_t pkg_count = 0;
-
-    while( !decoder->is_complete() )
-    {
-        encoder->encode( &payload[0] );
-        decoder->decode( &payload[0] );
-
-        ++pkg_count;
-    }
-
-    EXPECT_TRUE(pkg_count == encoder->symbols());
-
-    std::vector<uint8_t> data_out(decoder->block_size(), '\0');
-    decoder->copy_symbols(sak::storage(data_out));
-
-    EXPECT_TRUE(std::equal(data_out.begin(),
-                           data_out.end(),
-                           data_in.begin()));
-
-}
-
-
-
-
-
-
