+++ test/src/test_rlnc_full_vector_codes.cpp
@@ -22,7 +22,14 @@
 #include <kodo/has_shallow_symbol_storage.hpp>
 
 #include "basic_api_test_helper.hpp"
-#include "test_reuse.hpp"
+
+#include "helper_test_reuse_api.hpp"
+#include "helper_test_recoding_api.hpp"
+#include "helper_test_on_the_fly_api.hpp"
+#include "helper_test_basic_api.hpp"
+#include "helper_test_initialize_api.hpp"
+#include "helper_test_systematic_api.hpp"
+#include "helper_test_mix_uncoded_api.hpp"
 
 namespace kodo
 {
@@ -129,489 +136,98 @@ namespace kodo
 
 }
 
-template<template <class> class Encoder,
-         template <class> class Decoder>
-void test_coders(uint32_t symbols, uint32_t symbol_size)
-{
-    invoke_basic_api
-        <
-            Encoder<fifi::binary>,
-            Decoder<fifi::binary>
-            >(symbols, symbol_size);
-
-    invoke_basic_api
-        <
-            Encoder<fifi::binary8>,
-            Decoder<fifi::binary8>
-            >(symbols, symbol_size);
-
-    invoke_basic_api
-        <
-            Encoder<fifi::binary16>,
-            Decoder<fifi::binary16>
-            >(symbols, symbol_size);
-
-    invoke_basic_api
-        <
-            Encoder<fifi::prime2325>,
-            Decoder<fifi::prime2325>
-            >(symbols, symbol_size);
-}
-
-
-void test_coders(uint32_t symbols, uint32_t symbol_size)
-{
-
-    test_coders<
-        kodo::full_rlnc_encoder_shallow,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
-
-    test_coders<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
-
-    // The delayed decoders
-    test_coders<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed
-        >(symbols, symbol_size);
-}
 
 /// Tests the basic API functionality this mean basic encoding
 /// and decoding
-TEST(TestRlncFullVectorCodes, basic_api)
+TEST(TestRlncFullVectorCodes, test_basic_api)
 {
-    test_coders(32, 1600);
-    test_coders(1, 1600);
-
-    uint32_t symbols = rand_symbols();
-    uint32_t symbol_size = rand_symbol_size();
-
-    test_coders(symbols, symbol_size);
-}
-
-template
-    <
-    template <class> class Encoder,
-    template <class> class Decoder
-    >
-void test_initialize(uint32_t symbols, uint32_t symbol_size)
-{
-
-    invoke_initialize
-        <
-            Encoder<fifi::binary>,
-            Decoder<fifi::binary>
-            >(symbols, symbol_size);
-
-    invoke_initialize
-        <
-            Encoder<fifi::binary8>,
-            Decoder<fifi::binary8>
-            >(symbols, symbol_size);
-
-    invoke_initialize
-        <
-            Encoder<fifi::binary16>,
-            Decoder<fifi::binary16>
-            >(symbols, symbol_size);
-
-    // invoke_initialize
-    //     <
-    //         Encoder<fifi::prime2325>,
-    //         Decoder<fifi::prime2325>
-    //         >(symbols, symbol_size);
-
-}
-
-void test_initialize(uint32_t symbols, uint32_t symbol_size)
-{
-
-    test_initialize<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder>(symbols, symbol_size);
+    test_basic_api<kodo::full_rlnc_encoder_shallow, kodo::full_rlnc_decoder>();
+    test_basic_api<kodo::full_rlnc_encoder, kodo::full_rlnc_decoder>();
 
     // The delayed decoders
-    test_initialize<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed>(symbols, symbol_size);
-
+    test_basic_api<kodo::full_rlnc_encoder, kodo::full_rlnc_decoder_delayed>();
 }
+
+
 
 /// Test that the encoders and decoders initialize() function can be used
 /// to reset the state of an encoder and decoder and that they therefore
 /// can be safely reused.
-TEST(TestRlncFullVectorCodes, initialize_function)
+TEST(TestRlncFullVectorCodes, test_initialize)
 {
-    test_initialize(32, 1600);
-    test_initialize(1, 1600);
-
-    srand(static_cast<uint32_t>(time(0)));
-
-    uint32_t symbols = rand_symbols();
-    uint32_t symbol_size = rand_symbol_size();
-
-    test_initialize(symbols, symbol_size);
-}
-
-template<template <class> class Encoder,
-         template <class> class Decoder>
-void test_coders_systematic(uint32_t symbols, uint32_t symbol_size)
-{
-    invoke_systematic<
-        Encoder<fifi::binary>,
-        Decoder<fifi::binary>
-        >(symbols, symbol_size);
-
-    invoke_systematic<
-        Encoder<fifi::binary8>,
-        Decoder<fifi::binary8>
-        >(symbols, symbol_size);
-
-    invoke_systematic<
-        Encoder<fifi::binary16>,
-        Decoder<fifi::binary16>
-        >(symbols, symbol_size);
-
-    // invoke_systematic<
-    //     Encoder<fifi::prime2325>,
-    //     Decoder<fifi::prime2325>
-    //     >(symbols, symbol_size);
-
-}
-
-
-void test_coders_systematic(uint32_t symbols, uint32_t symbol_size)
-{
-    test_coders_systematic<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
-
+    test_initialize<kodo::full_rlnc_encoder_shallow, kodo::full_rlnc_decoder>();
+    test_initialize<kodo::full_rlnc_encoder, kodo::full_rlnc_decoder>();
 
     // The delayed decoders
-    test_coders_systematic<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed
-        >(symbols, symbol_size);
+    test_initialize<kodo::full_rlnc_encoder, kodo::full_rlnc_decoder_delayed>();
 
 }
 
 /// Tests that an encoder producing systematic packets is handled
 /// correctly in the decoder.
-TEST(TestRlncFullVectorCodes, systematic)
+TEST(TestRlncFullVectorCodes, test_systematic)
 {
-    test_coders_systematic(32, 1600);
-    test_coders_systematic(1, 1600);
-
-    srand(static_cast<uint32_t>(time(0)));
-
-    uint32_t symbols = rand_symbols();
-    uint32_t symbol_size = rand_symbol_size();
-
-    test_coders_systematic(symbols, symbol_size);
-}
-
-template<template <class> class Encoder,
-         template <class> class Decoder>
-void test_coders_raw(uint32_t symbols, uint32_t symbol_size)
-{
-    invoke_out_of_order_raw<
-        Encoder<fifi::binary>,
-        Decoder<fifi::binary>
-        >(symbols, symbol_size);
-
-    invoke_out_of_order_raw<
-        Encoder<fifi::binary8>,
-        Decoder<fifi::binary8>
-        >(symbols, symbol_size);
-
-    invoke_out_of_order_raw<
-        Encoder<fifi::binary16>,
-        Decoder<fifi::binary16>
-        >(symbols, symbol_size);
-
-    // invoke_out_of_order_raw<
-    //     Encoder<fifi::prime2325>,
-    //     Decoder<fifi::prime2325>
-    //     >(symbols, symbol_size);
-
-}
-
-void test_coders_raw(uint32_t symbols, uint32_t symbol_size)
-{
-    test_coders_raw<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
-
-    // The delayed decoders
-    test_coders_raw<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed
-        >(symbols, symbol_size);
+    test_systematic<kodo::full_rlnc_encoder, kodo::full_rlnc_decoder>();
+    test_systematic<kodo::full_rlnc_encoder, kodo::full_rlnc_decoder_delayed>();
 }
 
 /// Tests whether mixed un-coded and coded packets are correctly handled
 /// in the encoder and decoder.
-TEST(TestRlncFullVectorCodes, raw)
+TEST(TestRlncFullVectorCodes, mix_uncoded)
 {
-    test_coders_raw(32, 1600);
-    test_coders_raw(1, 1600);
-
-    uint32_t symbols = rand_symbols();
-    uint32_t symbol_size = rand_symbol_size();
-
-    test_coders_raw(symbols, symbol_size);
+    test_mix_uncoded<kodo::full_rlnc_encoder,kodo::full_rlnc_decoder>();
+    test_mix_uncoded<kodo::full_rlnc_encoder,kodo::full_rlnc_decoder_delayed>();
 }
 
-// Small helper structure holding the parameters needed for the
-// recoding tests.
-struct recoding_parameters
-{
-    uint32_t m_max_symbols;
-    uint32_t m_max_symbol_size;
-    uint32_t m_symbols;
-    uint32_t m_symbol_size;
-};
-
-//
-// Testing recoding
-//
-template<class Encoder, class Decoder>
-inline void invoke_recoding(recoding_parameters param)
+/// The recoding
+TEST(TestRlncFullVectorCodes, test_recoders_api)
 {
 
-    // Common setting
-    typename Encoder::factory encoder_factory(
-        param.m_max_symbols, param.m_max_symbol_size);
+    test_recoders<kodo::full_rlnc_encoder,
+        kodo::full_rlnc_decoder>();
 
-    encoder_factory.set_symbols(param.m_symbols);
-    encoder_factory.set_symbol_size(param.m_symbol_size);
+    test_recoders<kodo::full_rlnc_encoder,
+        kodo::full_rlnc_decoder_delayed>();
 
-    auto encoder = encoder_factory.build();
+    test_recoders<kodo::full_rlnc_encoder,
+        kodo::full_rlnc_decoder_delayed_shallow>();
 
-    typename Decoder::factory decoder_factory(
-        param.m_max_symbols, param.m_max_symbol_size);
-
-    decoder_factory.set_symbols(param.m_symbols);
-    decoder_factory.set_symbol_size(param.m_symbol_size);
-
-    auto decoder_one = decoder_factory.build();
-    auto decoder_two = decoder_factory.build();
-
-    // If tested with a shallow decoder we have to remember to set the
-    // buffers to use for the decoding
-    std::vector<uint8_t> buffer_decoder_one(decoder_one->block_size(), '\0');
-    std::vector<uint8_t> buffer_decoder_two(decoder_two->block_size(), '\0');
-
-    if(kodo::has_shallow_symbol_storage<Decoder>::value)
-    {
-        decoder_one->set_symbols(sak::storage(buffer_decoder_one));
-        decoder_two->set_symbols(sak::storage(buffer_decoder_two));
-    }
-
-    EXPECT_EQ(encoder->payload_size(), decoder_one->payload_size());
-    EXPECT_EQ(encoder->payload_size(), decoder_two->payload_size());
-
-    std::vector<uint8_t> payload(encoder->payload_size());
-    std::vector<uint8_t> data_in = random_vector(encoder->block_size());
-
-    encoder->set_symbols(sak::storage(data_in));
-
-    // Set the encoder non-systematic
-    if(kodo::is_systematic_encoder(encoder))
-        kodo::set_systematic_off(encoder);
-
-    while( !decoder_two->is_complete() )
-    {
-        uint32_t encode_size = encoder->encode( &payload[0] );
-        EXPECT_TRUE(encode_size <= payload.size());
-        EXPECT_TRUE(encode_size > 0);
-
-        decoder_one->decode( &payload[0] );
-
-        uint32_t recode_size = decoder_one->recode( &payload[0] );
-        EXPECT_TRUE(recode_size <= payload.size());
-        EXPECT_TRUE(recode_size > 0);
-
-        decoder_two->decode( &payload[0] );
-    }
-
-    std::vector<uint8_t> data_out_one(decoder_one->block_size(), '\0');
-    std::vector<uint8_t> data_out_two(decoder_two->block_size(), '\0');
-
-    decoder_one->copy_symbols(sak::storage(data_out_one));
-    decoder_two->copy_symbols(sak::storage(data_out_two));
-
-    EXPECT_TRUE(std::equal(data_out_one.begin(),
-                           data_out_one.end(),
-                           data_in.begin()));
-
-    EXPECT_TRUE(std::equal(data_out_two.begin(),
-                           data_out_two.end(),
-                           data_in.begin()));
-}
-
-
-template
-    <
-    template <class> class Encoder,
-    template <class> class Decoder
-    >
-void test_recoders(recoding_parameters param)
-{
-    invoke_recoding<
-        Encoder<fifi::binary>,
-        Decoder<fifi::binary> >(param);
-
-    invoke_recoding<
-        Encoder<fifi::binary8>,
-        Decoder<fifi::binary8> >(param);
-
-    invoke_recoding<
-        Encoder<fifi::binary16>,
-        Decoder<fifi::binary16> >(param);
-
-    // invoke_recoding<
-    //     Encoder<fifi::prime2325>,
-    //     Decoder<fifi::prime2325> >(
-    //     symbols, symbol_size);
-
-}
-
-void test_recoders(recoding_parameters param)
-{
-
-    test_recoders
-        <
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder>(param);
-
-    test_recoders<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed>(param);
-
-    test_recoders<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed_shallow>(param);
-
-}
-
-/// Tests that the recoding function works, this is done by using one
-/// encoder
-/// and two decoders:
-///
-///    +------------+      +------------+      +------------+
-///    | encoder    |+---->| decoder #1 |+---->| decoder #2 |
-///    +------------+      +------------+      +------------+
-///
-/// Where the encoder passes data to the first decoder which then
-/// recodes and passes data to the second decoder
-///
-TEST(TestRlncFullVectorCodes, recoding_simple)
-{
-    recoding_parameters param;
-    param.m_max_symbols = 32;
-    param.m_max_symbol_size = 1600;
-    param.m_symbols = param.m_max_symbols;
-    param.m_symbol_size = param.m_max_symbol_size;
-
-    test_recoders(param);
-
-    param.m_max_symbols = 1;
-    param.m_max_symbol_size = 1600;
-    param.m_symbols = param.m_max_symbols;
-    param.m_symbol_size = param.m_max_symbol_size;
-
-
-    test_recoders(param);
-
-    param.m_max_symbols = 8;
-    param.m_max_symbol_size = 8;
-    param.m_symbols = param.m_max_symbols;
-    param.m_symbol_size = param.m_max_symbol_size;
-
-    test_recoders(param);
-
-    param.m_max_symbols = rand_symbols();
-    param.m_max_symbol_size = rand_symbol_size();
-    param.m_symbols = rand_symbols(param.m_max_symbols);
-    param.m_symbol_size = rand_symbol_size(param.m_max_symbol_size);
-
-    test_recoders(param);
-}
-
-
-
-
-void test_reuse(uint32_t symbols, uint32_t symbol_size)
-{
-
-    test_reuse<
-        kodo::full_rlnc_encoder_shallow,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
-
-    test_reuse<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
-
-    // The delayed decoders
-    test_reuse<
-        kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed
-        >(symbols, symbol_size);
 }
 
 /// Tests the basic API functionality this mean basic encoding
 /// and decoding
-TEST(TestRlncFullVectorCodes, test_reuse)
+TEST(TestRlncFullVectorCodes, test_reuse_api)
 {
-    test_reuse(32, 1600);
-    test_reuse(1, 1600);
+    test_reuse<
+        kodo::full_rlnc_encoder_shallow,
+        kodo::full_rlnc_decoder>();
 
-    uint32_t symbols = rand_symbols();
-    uint32_t symbol_size = rand_symbol_size();
+    test_reuse<
+        kodo::full_rlnc_encoder,
+        kodo::full_rlnc_decoder>();
 
-    test_reuse(symbols, symbol_size);
+    // The delayed decoders
+    test_reuse<
+        kodo::full_rlnc_encoder,
+        kodo::full_rlnc_decoder_delayed>();
 }
 
-void test_reuse_incomplete(uint32_t symbols, uint32_t symbol_size)
+/// Tests the basic API functionality this mean basic encoding
+/// and decoding
+TEST(TestRlncFullVectorCodes, test_reuse_incomplete_api)
 {
-
     test_reuse_incomplete<
         kodo::full_rlnc_encoder_shallow,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
+        kodo::full_rlnc_decoder>();
 
     test_reuse_incomplete<
         kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder
-        >(symbols, symbol_size);
+        kodo::full_rlnc_decoder>();
 
     // The delayed decoders
     test_reuse_incomplete<
         kodo::full_rlnc_encoder,
-        kodo::full_rlnc_decoder_delayed
-        >(symbols, symbol_size);
-}
-
-
-/// Tests the basic API functionality this mean basic encoding
-/// and decoding
-TEST(TestRlncFullVectorCodes, test_reuse_incomplete)
-{
-    test_reuse(32, 1600);
-    test_reuse(1, 1600);
-
-    uint32_t symbols = rand_symbols();
-    uint32_t symbol_size = rand_symbol_size();
-
-    test_reuse_incomplete(symbols, symbol_size);
+        kodo::full_rlnc_decoder_delayed>();
 }
