+++ test/src/test_rs_reed_solomon_codes.cpp
@@ -10,6 +10,7 @@
 
 #include "basic_api_test_helper.hpp"
 
+
 TEST(TestReedSolomonCodes, test_construct)
 {
